move _xml_from_anything to core.util
